+++ src/controllers/subject.controller.ts
@@ -1,5 +1,4 @@
-import { Request, Response, CookieOptions } from "express";
-import mongoose from "mongoose";
+import { Response } from "express";
 import AsyncHandler from "../utils/AsyncHandler";
 import ApiError from "../utils/ApiError";
 import ApiResponse from "../utils/ApiResponse";
@@ -8,29 +7,29 @@ import { AuthenticatedRequest } from "../middlewares/auth.middleware";
 
 
 const getSubjects = AsyncHandler(async (req: AuthenticatedRequest, res: Response) => {
-    const userId = (req as any).user?._id?.toString();
-    const subjects = await Subject.find({ owner: userId });
+    if (!req.user) throw new ApiError('Unauthorized', 401);
+    const userId = req.user._id;
+    const subjects = await Subject.find({ owner: userId as any });
 
     return res.status(200).json(new ApiResponse(true, 'Subjects fetched', subjects));
 });
 
 const addSubject = AsyncHandler(async (req: AuthenticatedRequest, res: Response) => {
     const { subName, desc } = req.body;
-    const userId = (req as any).user?._id; 
+    if (!req.user) throw new ApiError('Unauthorized', 401);
+    const userId = req.user._id; 
 
     if (!subName) throw new ApiError("Could not get subject name", 400);
 
-    const subExist = await Subject.findOne({ subName, owner: userId });
+    const subExist = await Subject.findOne({ subName, owner: userId as any });
     if (subExist)
         return res.status(400).json(new ApiResponse(false, "Subject already exists", subExist));
 
     const subject = await Subject.create({
         subName,
         desc,
-        owner: userId, 
+        owner: userId as any, 
     });
-
-    // console.log("Saved subject owner:", subject.owner);
 
     return res.status(201).json(new ApiResponse(true, "New Subject Added", subject));
 });
@@ -39,21 +38,24 @@ const addSubject = AsyncHandler(async (req: AuthenticatedRequest, res: Response)
 
 const editSubject = AsyncHandler(async(req:AuthenticatedRequest,res:Response)=>{
     const {subName,desc} = req.body;
-
-    const userId = (req as any).user?._id?.toString();
+    if (!req.user) throw new ApiError('Unauthorized', 401);
+    const userId = req.user._id;
 
     if(!subName||!desc) throw new ApiError("Could not get subject name or description",402)
 
     const {id} = req.params;
 
-    const subject = await Subject.findByIdAndUpdate(
-        { _id: id, owner: userId },
+    const subject = await Subject.findOneAndUpdate(
+        { _id: id, owner: userId as any },
         {
             subName,
             desc,
         },
         { new: true }
     );
+    
+    if (!subject) throw new ApiError("Subject not found or unauthorized", 404);
+
     return res
     .status(200)
     .json(new ApiResponse(true,"Subject Updated",subject));
@@ -61,14 +63,17 @@ const editSubject = AsyncHandler(async(req:AuthenticatedRequest,res:Response)=>{
 
 const delSubject = AsyncHandler(async(req:AuthenticatedRequest,res:Response)=>{
     const {id:subId} = req.params
+    if (!req.user) throw new ApiError('Unauthorized', 401);
+    const userId = req.user._id;
 
     await Subject.deleteOne({
-        _id:subId
+        _id: subId,
+        owner: userId as any
     });
 
     return res
     .status(200)
-    .json(new ApiResponse(true,"Subject Subject Deleted",null));
+    .json(new ApiResponse(true,"Subject Deleted",null));
 })
 
 export {
@@ -77,4 +82,3 @@ export {
     editSubject,
     delSubject
 }
-
